examples/540-livekit-agents-python/requirements.txt
# LiveKit Agents framework
livekit-agents>=1.5.0

# Deepgram plugin for LiveKit Agents
livekit-plugins-deepgram>=1.5.0

# OpenAI plugin for LLM (GPT-4o-mini)
livekit-plugins-openai>=1.5.0

# Environment variable management
python-dotenv>=1.0.0
